// src/api/models/GetProviders200Response.ts
export interface GetProviders200Response {
    /**
     * 
     * @type {number}
     * @memberof GetProviders200Response
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof GetProviders200Response
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof GetProviders200Response
     */
    messageEn: string;
    /**
     * 
     * @type {string}
     * @memberof GetProviders200Response
     */
    messageEs: string;
    /**
     * 
     * @type {string}
     * @memberof GetProviders200Response
     */
    messageDe: string;
    /**
     * 
     * @type {string}
     * @memberof GetProviders200Response
     */
    logo: string;
    /**
     * 
     * @type {GetProviders200ResponseStyle}
     * @memberof GetProviders200Response
     */
    style: GetProviders200ResponseStyle;
    /**
     * 
     * @type {GetProviders200ResponseCta}
     * @memberof GetProviders200Response
     */
    cta: GetProviders200ResponseCta;
    /**
     * Create
     * @type {any}
     * @memberof GetProviders200Response
     */
    _201?: any | null;
    /**
     * Not Found
     * @type {any}
     * @memberof GetProviders200Response
     */
    _404?: any | null;
}